feat: add deterministic local CSS naming option
This feature introduces a new option `DeterministicLocalCSSNaming` that, when enabled, ensures that
local CSS class names are generated in a deterministic manner. This is particularly useful for
caching and consistent builds across different environments.

When `MinifyIdentifiers` is enabled, the local name is generated as the original CSS class name, and
a hash of the absolute source file path, separated by an underscore (eg.
`<original-class-name>_<hash-of-absolute-path>`).

When `MinifyIdentifiers` is disabled, the local name is appended with a CSS safe version of the
relative source file path (eg.
`<original-class-name>_<hash-of-absolute-path>_<relative-source-file-path>`).